Make the bodyElement more lazy and aggressive.
It will now only set the body element if someone tries to access it and
will attempt to set it if it is ever null.

Previously setting in on ready caused an unknown order of execution with other methods
that might depend on this and it could be null in some cases, now it will
aggressively try to set it any time its null preventing this ordering issue.